# tp: Reduce the number of trace_bounds calculations.