@LiangliangSui 2025-07-03
> If we want to improve this we need a more targeted change where we detect whether or not the source tables have changed and decide based on that whether or not we should do this.

https://github.com/google/perfetto/pull/2022 only changed the three calls to BuildBoundsTable into two. Maybe really necessary to detect the changes of tables to decide whether to call BuildBoundsTable.