@LalitMaganti 2025-07-03
> https://github.com/google/perfetto/pull/2022 only changed the three calls to BuildBoundsTable into two. Maybe really necessary to detect the changes of tables to decide whether to call BuildBoundsTable.

Can you explain where the two are coming from? There should only be a single call.