@LiangliangSui 2025-07-03
Flush and below link
https://github.com/google/perfetto/blob/05523df14a57fdada791d362968e4730a86d8633/src/trace_processor/trace_processor_impl.cc#L623-L629