@LalitMaganti 2025-07-03
If you read the change caarefully, Flush is no longer called by trace_processor_shell when not symbolizing/deobfuscating. I'm very sure if you test your traces with that patch, you will see the same perf improvement as your patch.