@LiangliangSui 2025-07-03
sorry, I didn't notice the changes in src/trace_processor/trace_processor_shell.cc. The current modifications can achieve the expected results.