@LiangliangSui 2025-07-03
I have a question unrelated to this PR.

I want to build a tool to split a large trace into multiple small traces. Could you share your thoughts or suggestions?